Move ResultSet to Cassandra hierarchy
