+++ services/galley/src/Galley/Cassandra/ResultSet.hs
@@ -15,7 +15,7 @@
 -- You should have received a copy of the GNU Affero General Public License along
 -- with this program. If not, see <https://www.gnu.org/licenses/>.
 
-module Galley.Data.ResultSet where
+module Galley.Cassandra.ResultSet where
 
 import Cassandra
 import Imports
